[FIX] Section ID is a long
https://github.com/cloudbees/zendesk-java-client/blob/master/src/main/java/org/zendesk/client/v2/model/hc/Section.java#L10
https://developer.zendesk.com/rest_api/docs/help_center/sections#json-format (marked as int but they aren't  - it is the same for others objects)

+++ src/main/java/org/zendesk/client/v2/Zendesk.java
@@ -1631,7 +1631,7 @@ public class Zendesk implements Closeable {
             handleList(Section.class, "sections"));
     }
 
-    public Section getSection(int id) {
+    public Section getSection(long id) {
         return complete(submit(req("GET", tmpl("/help_center/sections/{id}.json").set("id", id)),
                 handle(Section.class, "section")));
     }
